+++ src/main/java/iudx/catalogue/server/database/rating/CreateRating.java
@@ -1,0 +1,84 @@
+package iudx.catalogue.server.database.rating;
+
+import static iudx.catalogue.server.database.Constants.GET_RDOC_QUERY;
+import static iudx.catalogue.server.util.Constants.*;
+
+import io.vertx.core.AsyncResult;
+import io.vertx.core.Future;
+import io.vertx.core.Handler;
+import io.vertx.core.json.JsonObject;
+import iudx.catalogue.server.database.ElasticClient;
+import iudx.catalogue.server.database.RespBuilder;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
+public class CreateRating {
+  private static final Logger LOGGER = LogManager.getLogger(CreateRating.class);
+  ElasticClient client;
+  String ratingIndex;
+
+  public CreateRating(ElasticClient client, String ratingIndex) {
+    this.client = client;
+    this.ratingIndex = ratingIndex;
+  }
+
+  public void createRating(JsonObject ratingDoc, Handler<AsyncResult<JsonObject>> handler) {
+    RespBuilder respBuilder = new RespBuilder();
+    String ratingId = ratingDoc.getString("ratingID");
+
+    String checkForExistingRecord = GET_RDOC_QUERY.replace("$1", ratingId).replace("$2", "");
+
+    client.searchAsync(
+        checkForExistingRecord,
+        ratingIndex,
+        checkRes -> {
+          if (checkRes.failed()) {
+            LOGGER.error("Fail: Insertion of rating failed: " + checkRes.cause());
+            handler.handle(
+                Future.failedFuture(
+                    respBuilder
+                        .withType(FAILED)
+                        .withResult(ratingId)
+                        .withDetail("Fail: Insertion of rating failed")
+                        .getResponse()));
+          } else {
+            if (checkRes.result().getInteger(TOTAL_HITS) != 0) {
+              handler.handle(
+                  Future.failedFuture(
+                      respBuilder
+                          .withType(TYPE_ALREADY_EXISTS)
+                          .withTitle(TITLE_ALREADY_EXISTS)
+                          .withResult(ratingId, INSERT, FAILED, " Fail: Doc Already Exists")
+                          .withDetail(" Fail: Doc Already Exists")
+                          .getResponse()));
+              return;
+            }
+
+            client.docPostAsync(
+                ratingIndex,
+                ratingDoc.toString(),
+                postRes -> {
+                  if (postRes.succeeded()) {
+                    handler.handle(
+                        Future.succeededFuture(
+                            respBuilder
+                                .withType(TYPE_SUCCESS)
+                                .withTitle(TITLE_SUCCESS)
+                                .withResult(ratingId, INSERT, TYPE_SUCCESS)
+                                .getJsonResponse()));
+                  } else {
+
+                    handler.handle(
+                        Future.failedFuture(
+                            respBuilder
+                                .withType(TYPE_FAIL)
+                                .withResult(ratingId, INSERT, FAILED)
+                                .withDetail("Insertion Failed")
+                                .getResponse()));
+                    LOGGER.error("Fail: Insertion failed" + postRes.cause());
+                  }
+                });
+          }
+        });
+  }
+}
