+++ src/main/java/iudx/catalogue/server/database/rating/DeleteRating.java
@@ -1,0 +1,83 @@
+package iudx.catalogue.server.database.rating;
+
+import static iudx.catalogue.server.database.Constants.GET_RDOC_QUERY;
+import static iudx.catalogue.server.util.Constants.*;
+import static iudx.catalogue.server.util.Constants.TITLE_SUCCESS;
+
+import io.vertx.core.AsyncResult;
+import io.vertx.core.Future;
+import io.vertx.core.Handler;
+import io.vertx.core.json.JsonObject;
+import iudx.catalogue.server.database.ElasticClient;
+import iudx.catalogue.server.database.RespBuilder;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
+public class DeleteRating {
+  private static final Logger LOGGER = LogManager.getLogger(DeleteRating.class);
+  private static final String internalErrorResp =
+      new RespBuilder()
+          .withType(TYPE_INTERNAL_SERVER_ERROR)
+          .withTitle(TITLE_INTERNAL_SERVER_ERROR)
+          .withDetail(DETAIL_INTERNAL_SERVER_ERROR)
+          .getResponse();
+  ElasticClient client;
+  String ratingIndex;
+
+  public DeleteRating(ElasticClient client, String ratingIndex) {
+    this.client = client;
+    this.ratingIndex = ratingIndex;
+  }
+
+  public void deleteRating(JsonObject request, Handler<AsyncResult<JsonObject>> handler) {
+
+    RespBuilder respBuilder = new RespBuilder();
+    String ratingId = request.getString("ratingID");
+
+    String checkForExistingRecord = GET_RDOC_QUERY.replace("$1", ratingId).replace("$2", "");
+
+    client.searchGetId(
+        checkForExistingRecord,
+        ratingIndex,
+        checkRes -> {
+          if (checkRes.failed()) {
+            LOGGER.error("Fail: Check query fail;" + checkRes.cause());
+            handler.handle(Future.failedFuture(internalErrorResp));
+          } else {
+            if (checkRes.result().getInteger(TOTAL_HITS) != 1) {
+              LOGGER.error("Fail: Doc doesn't exist, can't delete");
+              handler.handle(
+                  Future.failedFuture(
+                      respBuilder
+                          .withType(TYPE_ITEM_NOT_FOUND)
+                          .withTitle(TITLE_ITEM_NOT_FOUND)
+                          .withResult(
+                              ratingId, DELETE, FAILED, "Fail: Doc doesn't exist, can't delete")
+                          .withDetail("Fail: Doc doesn't exist, can't delete")
+                          .getResponse()));
+              return;
+            }
+
+            String docId = checkRes.result().getJsonArray(RESULTS).getString(0);
+
+            client.docDelAsync(
+                docId,
+                ratingIndex,
+                putRes -> {
+                  if (putRes.succeeded()) {
+                    handler.handle(
+                        Future.succeededFuture(
+                            respBuilder
+                                .withType(TYPE_SUCCESS)
+                                .withTitle(TITLE_SUCCESS)
+                                .withResult(ratingId)
+                                .getJsonResponse()));
+                  } else {
+                    handler.handle(Future.failedFuture(internalErrorResp));
+                    LOGGER.error("Fail: Deletion failed;" + putRes.cause());
+                  }
+                });
+          }
+        });
+  }
+}
